Update the tasks, notes and logs in the readme
I've added the Browser Jab submission to logs and added log's from yesterday and today.

I've also polished the notes to be even in terms of focus and combined some.

Lastly, I have updated the "roadmap" in form of the tasks to look at next with the ideas I currently have.